remove models and admin to payment
